Bugs fixed
